move conv outside of util
